requirements.txt
flask==3.0.0
numpy==1.24.3
tensorflow==2.15.0
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78
redis==4.6.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
requests==2.31.0